requirements.txt
asgiref==3.5.2
astor==0.8.1
attrdict==2.0.1
Babel==2.10.3
bce-python-sdk==0.8.74
beautifulsoup4==4.11.1
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cssselect==1.1.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
einops==0.2.0
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.2.2
Flask-Babel==2.0.0
fonttools==4.37.1
future==0.18.2
gdown==4.4.0
idna==3.3
imageio==2.21.2
imgaug==0.4.0
itsdangerous==2.1.2
jarowinkler==1.2.1
Jinja2==3.1.2
kiwisolver==1.4.4
lmdb==1.0.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
networkx==2.8.6
numpy==1.23.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddlepaddle==2.3.2
pandas==1.4.4
Pillow==9.2.0
prefetch-generator==1.0.1
premailer==3.10.0
protobuf==3.20.0
pyclipper==1.3.0.post3
pycryptodome==3.15.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
rapidfuzz==2.6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.1
Shapely==1.8.4
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tifffile==2022.8.12
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
typing_extensions==4.3.0
tzdata==2022.2
Unidecode==1.3.4
urllib3==1.26.12
vietocr==0.3.8
visualdl==2.4.0
Werkzeug==2.2.2
wincertstore==0.2
